Refactor path handling functionality to separate helper class
Refs #368
